Rename the components
